requirements.txt
# Unified requirements for VH-RAD Traffic Monitor
# This combines minimal and enhanced requirements

# Core dependencies
requests>=2.31.0
python-dateutil>=2.8.2
pydantic>=2.5.0
pydantic-settings>=2.0.0
PyYAML>=6.0.1

# FastAPI and server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
websockets>=12.0
httpx>=0.27.0
python-multipart>=0.0.9

# Production enhancements
slowapi>=0.1.9
pybreaker>=1.0.1
python-json-logger>=2.0.7
structlog>=24.1.0

# Redis for caching (recommended for production)
redis>=5.0.0
aioredis>=2.0.1

# Prometheus metrics for monitoring
prometheus-client>=0.19.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Elasticsearch integration (optional)
# elasticsearch>=8.0.0

# Development and testing (optional)
# Install with: pip install -r requirements.txt -r requirements-dev.txt
# pytest>=8.0.0
# pytest-asyncio>=0.23.0
# pytest-cov>=5.0.0
# pytest-mock>=3.14.0

# Performance enhancements (optional)
# orjson>=3.9.0
# ujson>=5.9.0
